Tighten runtime report and lint contracts
Drop legacy runtime compatibility paths, refine report request contracts, and promote the current Ruff rule backlog into the default static gate.

Retarget strict Ruff preview to PTH201 while preserving filesystem transaction and public-python path resolution behavior with focused tests.

+++ ops/scripts/core/observability_routing_provenance_runtime.py
@@ -462,7 +462,7 @@ def write_routing_provenance_aggregate(
         phase="routing_provenance_aggregate",
         decision="written",
         artifact_path=written_rel_path,
-        duration_ms=int(round((context.utcnow() - started_at).total_seconds() * 1000)),
+        duration_ms=round((context.utcnow() - started_at).total_seconds() * 1000),
         session_id=session_id,
         policy_version=dict(session.get("policy", {})).get("version", ""),
     )
